Broke down EventStoreDB extensions into dedicated files (configuration and operations)
