+++ Sample/EventStoreDB/Simple/ECommerce.Core/Configuration.cs
@@ -1,4 +1,5 @@
 ﻿using ECommerce.Core.Events;
+using ECommerce.Core.EventStoreDB;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
