+++ Sample/EventStoreDB/Simple/ECommerce.Core/EventStoreDB/EventStoreDBExtensions.cs
@@ -1,25 +1,13 @@
-﻿using System;
+using System;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 using ECommerce.Core.Serialisation;
-using ECommerce.Core.Subscriptions;
 using EventStore.Client;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.DependencyInjection;
 
-namespace ECommerce.Core
+namespace ECommerce.Core.EventStoreDB
 {
-    public class EventStoreDBConfig
-    {
-        public string ConnectionString { get; set; } = default!;
-    }
-
-    public record EventStoreDBOptions(
-        bool UseInternalCheckpointing = true
-    );
-
-    public static class EventStoreDBConfigExtensions
+    public static class EventStoreDBExtensions
     {
         public static async Task<TEntity> Find<TEntity>(
             this EventStoreClient eventStore,
@@ -75,26 +63,6 @@ namespace ECommerce.Core
                 new[] { @event.ToJsonEventData() },
                 cancellationToken: cancellationToken
             );
-        }
-
-
-        private const string DefaultConfigKey = "EventStore";
-
-        public static IServiceCollection AddEventStoreDB(this IServiceCollection services, IConfiguration config, EventStoreDBOptions? options = null)
-        {
-            var eventStoreDBConfig = config.GetSection(DefaultConfigKey).Get<EventStoreDBConfig>();
-
-            services.AddSingleton(
-                    new EventStoreClient(EventStoreClientSettings.Create(eventStoreDBConfig.ConnectionString)))
-                .AddTransient<EventStoreDBSubscriptionToAll, EventStoreDBSubscriptionToAll>();
-
-            if (options?.UseInternalCheckpointing != false)
-            {
-                services
-                    .AddTransient<ISubscriptionCheckpointRepository, EventStoreDBSubscriptionCheckpointRepository>();
-            }
-
-            return services;
         }
     }
 }
